requirements.txt
mysql-python
django-allauth
pillow
django-widget-tweaks
